Move admin components into 'provincial-admin' folder
