+++ embc-app/ClientApp/src/app/app-routing.module.ts
@@ -24,16 +24,16 @@ import { RedirectGuard } from './core/guards/redirect.guard';
 import { VOLUNTEER, LOCAL_AUTHORITY, PROVINCIAL_ADMIN } from './constants';
 import { LandingPageGuard } from './core/guards/landing-page.guard';
 import { UsefulInformationContentComponent } from './useful-information-content/useful-information-content.component';
-import { TaskNumberListComponent } from './task-number-list/task-number-list.component';
+import { TaskNumberListComponent } from './provincial-admin/components/task-number-list/task-number-list.component';
 import { VolunteerListComponent } from './volunteer-list/volunteer-list.component';
-import { OrganizationListComponent } from './organization-list/organization-list.component';
-import { OrganizationMakerComponent } from './organization-maker/organization-maker.component';
-import { TaskNumberMakerComponent } from './task-number-maker/task-number-maker.component';
+import { OrganizationListComponent } from './provincial-admin/components/organization-list/organization-list.component';
+import { OrganizationMakerComponent } from './provincial-admin/components/organization-maker/organization-maker.component';
+import { TaskNumberMakerComponent } from './provincial-admin/components/task-number-maker/task-number-maker.component';
 import { VolunteerMakerComponent } from './volunteer-maker/volunteer-maker.component';
 import { RegistrationListComponent } from './registration-list/registration-list.component';
 import { VolunteerUsefulInformationComponent } from './volunteer-useful-information/volunteer-useful-information.component';
 import { RegistrationSummaryFullComponent } from './registration-summary-full/registration-summary-full.component';
-import { VolunteerOrganizationListComponent } from './volunteer-organization-list/volunteer-organization-list.component';
+import { VolunteerOrganizationListComponent } from './provincial-admin/components/volunteer-organization-list/volunteer-organization-list.component';
 import { SessionExpiredComponent } from './session-expired/session-expired.component';
 import { VolunteerRegistrationsPageComponent } from './pages/volunteer/volunteer-registrations-page/volunteer-registrations-page.component';
 import { UsefulInformationPageComponent } from './pages/useful-information-page/useful-information-page.component';
@@ -42,6 +42,7 @@ import { ProvincialAdminOrganizationsPageComponent } from './provincial-admin/pa
 import { ProvincialAdminTaskNumbersPageComponent } from './provincial-admin/pages/task-numbers/provincial-admin-task-numbers-page.component';
 import { ProvincialAdminRegistrationsPageComponent } from './provincial-admin/pages/registrations/provincial-admin-registrations-page.component';
 import { LocalAuthorityVolunteersPageComponent } from './pages/local-authority/local-authority-volunteers-page/local-authority-volunteers-page.component';
+import { ProvincialAdminVolunteersOrganizationPageComponent } from './provincial-admin/pages/volunteers-organization/provincial-admin-volunteers-organization-page.component';
 
 /*
 
@@ -277,6 +278,11 @@ const routes: Routes = [
         data: { expectedRole: PROVINCIAL_ADMIN },
       },
       {
+        path: 'organization/volunteers',
+        component: ProvincialAdminVolunteersOrganizationPageComponent,
+        data: { expectedRole: PROVINCIAL_ADMIN },
+      },
+      {
         path: 'task-numbers',
         component: ProvincialAdminTaskNumbersPageComponent,
         data: { expectedRole: PROVINCIAL_ADMIN },
@@ -299,11 +305,6 @@ const routes: Routes = [
       {
         path: 'organization',
         component: OrganizationMakerComponent,
-        data: { expectedRole: PROVINCIAL_ADMIN },
-      },
-      {
-        path: 'organization/volunteers',
-        component: VolunteerOrganizationListComponent,
         data: { expectedRole: PROVINCIAL_ADMIN },
       },
       {
