+++ embc-app/ClientApp/src/app/app.module.ts
@@ -21,17 +21,17 @@ import { VolunteerTeamDashboardComponent } from './volunteer-team-dashboard/volu
 import { VolunteerUsefulInformationComponent } from './volunteer-useful-information/volunteer-useful-information.component';
 import { RegistrationSummaryComponent } from './registration-summary/registration-summary.component';
 import { TopNavigationComponent } from './top-navigation/top-navigation.component';
-import { TaskNumberListComponent } from './task-number-list/task-number-list.component';
+import { TaskNumberListComponent } from './provincial-admin/components/task-number-list/task-number-list.component';
 import { VolunteerListComponent } from './volunteer-list/volunteer-list.component';
 import { RegistrationListComponent } from './registration-list/registration-list.component';
-import { OrganizationListComponent } from './organization-list/organization-list.component';
+import { OrganizationListComponent } from './provincial-admin/components/organization-list/organization-list.component';
 import { UsefulInformationContentComponent } from './useful-information-content/useful-information-content.component';
 import { VolunteerEditorContainerComponent } from './volunteer-editor-container/volunteer-editor-container.component';
-import { TaskNumberMakerComponent } from './task-number-maker/task-number-maker.component';
+import { TaskNumberMakerComponent } from './provincial-admin/components/task-number-maker/task-number-maker.component';
 import { VolunteerMakerComponent } from './volunteer-maker/volunteer-maker.component';
-import { OrganizationMakerComponent } from './organization-maker/organization-maker.component';
+import { OrganizationMakerComponent } from './provincial-admin/components/organization-maker/organization-maker.component';
 import { RegistrationSummaryFullComponent } from './registration-summary-full/registration-summary-full.component';
-import { VolunteerOrganizationListComponent } from './volunteer-organization-list/volunteer-organization-list.component';
+import { VolunteerOrganizationListComponent } from './provincial-admin/components/volunteer-organization-list/volunteer-organization-list.component';
 import { LayoutMainComponent } from './components/layout-main/layout-main.component';
 import { LayoutSideComponent } from './components/layout-side/layout-side.component';
 import { RegistrationMakerComponent } from './registration-maker/registration-maker.component';
