+++ embc-app/ClientApp/src/app/provincial-admin/components/organization-list/organization-list.component.html
@@ -67,7 +67,7 @@
         {{notFoundMessage}}
       </section>
       <section>
-        <button class="btn btn-primary" (click)="modifyOrganization(organization.id)">ADD ORGANIZATION</button>
+        <button class="btn btn-primary" (click)="modifyOrganization()">ADD ORGANIZATION</button>
       </section>
     </div>
   </ng-template>
